service-auth/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
python-dotenv
sqlalchemy
psycopg2-binary
pydantic-settings
httpx
requests
python-jose[cryptography]
passlib[bcrypt]
bcrypt
pytest
pytest-cov
pytest-asyncio
pytest-mock
huggingface_hub
pydantic_settings
jwt
email-validator